backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-google-genai
langchain_anthropic
langchain-cohere
langchain-deepseek
python-dotenv
grpcio  